TST: Move Hypothesis strategies into own file
